You know what? Fuck UPS and their stupid ass business rules.
It's a fucking API. I pass you data, you take care of the business
rules.  "Required if shipper is in the US or CA. If Shipper country is
US or CA, then the value must be a valid US State/ Canadian Province
code. If the country is Ireland, the StateProvinceCode will contain the
county."  Fuck yourself. If I pass you something in that field and it's
being sent from not the US, CA, or IE just fucking ignore it. Don't tell
me it's invalid.